Create email.py
This takes an array of e-mail addresses, and sends an e-mail to each, while also including the e-mail of each user in the body. This, allows for an individual user to receive a personalized e-mail, where their e-mail is the username, but not all e-mails are in the same formatting. Thus, giving a unique result to each user.